requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
python-dotenv
email-validator
Flask-Bootstrap
gunicorn
google-generativeai